+++ tf_encrypted/protocol/securenn/__init__.py
@@ -1,0 +1,7 @@
+from __future__ import absolute_import
+
+from .securenn import SecureNN
+
+__all__ = [
+    "SecureNN",
+]
