+++ tf_encrypted/tensor/__init__.py
@@ -1,27 +1,26 @@
 from __future__ import absolute_import
 
-from .odd_implicit import oddInt64factory
-
 from .int100 import (
     int100factory,
     Int100Constant,
     Int100Placeholder,
     Int100Variable,
-    Int100Tensor
+    Int100Tensor,
 )
 
 from .prime import (
     PrimeTensor,
     PrimePlaceholder,
     PrimeVariable,
-    PrimeConstant
+    PrimeConstant,
+    PrimeFactory,
 )
 
 from .int32 import (
     Int32Tensor,
     Int32Placeholder,
     Int32Variable,
-    Int32Constant
+    Int32Constant,
 )
 
 from .int64 import (
@@ -29,7 +28,7 @@ from .int64 import (
     Int64Tensor,
     Int64Placeholder,
     Int64Variable,
-    Int64Constant
+    Int64Constant,
 )
 
 from .fixed import (
@@ -54,6 +53,7 @@ __all__ = [
     'PrimePlaceholder',
     'PrimeVariable',
     'PrimeConstant',
+    'PrimeFactory',
     'Int32Tensor',
     'Int32Placeholder',
     'Int32Variable',
@@ -62,5 +62,4 @@ __all__ = [
     'Int64Placeholder',
     'Int64Variable',
     'Int64Constant',
-    'oddInt64factory',
 ]
